Add --rm-before-starting option
The get_ipiot_source.sh script does not work from a git checkout;
it tries to overwrite files that already exist.  Instead of failing
weirdly, we now fail-fast and offer the --rm-before-starting option
that nukes the prior files and then rebuilds the tree.  Other than
the failure checking and recovery, this is a non-functional change;
it does not alter the resulting ipopt-mirror results at all.