Pin Ipopt download to a tag instead of a branch.
The stable/3.12 is actually a long-lived svn release branch, not a
fixed-in-time tag.  We should use a tag for reproducability, which
Ipopt calls, e.g., release/3.12.5.

Even worse, the long-lived release branch actually refers to the
long-lived release branches of the ThirdParty items!  So even if we
knew which revision of Ipopt we'd happened to pull off of the release
branch, we'd need date-based commit magic to pull the ThirdParty
versions that we used.

Insetad of doing all that, we just re-pin to 3.12.5 here and eat
the (relatively minor) perturbations that result.